Refactor: NeighborReply 위치 변경
