+++ src/main/kotlin/com/wafflestudio/team03server/core/neighbor/entity/NeighborComment.kt
@@ -1,7 +1,6 @@
 package com.wafflestudio.team03server.core.neighbor.entity
 
 import com.wafflestudio.team03server.common.BaseTimeEntity
-import com.wafflestudio.team03server.core.neighbor.api.response.NeighborReply
 import com.wafflestudio.team03server.core.user.entity.User
 import javax.persistence.*
